low and high memory confs
